Refactor and centralize utility logic (issue #14)
Separate utility functions and business logic from bot orchestration code
for better maintainability and testability.

Changes:
- Created filters.py for AE detection logic
  - Moved AE_TRIGGERS constant
  - Moved is_arbitration_enforcement() function

- Enhanced timestamp.py with additional utilities
  - Moved clean_invisible_unicode() function
  - Moved format_expiry() function
  - Updated module docstring to reflect broader scope

- Updated bot.py to import from new locations
  - Removed duplicate function definitions
  - Removed unused unicodedata import
  - Updated imports from filters and timestamp modules

- Updated test imports to reflect new module structure

This separation makes the code more modular and easier to test in isolation.
